Code spike for retrieving javadocs from source files
A spike to experiment with retrieving (building) javadocs from source
files. The advantage to this approach is it would support documentation
from 3rd party libraries. The challenge is all of this needs to be built
manually the javadocs do all of the assembling of information
@inheritDoc, methods, parent, sibling, children navigation, etc.

It is possible but becomes very difficult to read and reason about. If
the JDK html javadoc retrieval becomes too limiting this is a valid
alternative path.